changed short to int

+++ TimeTool/src/main/java/persistency/year/Year.java
@@ -7,7 +7,7 @@ import java.util.TreeMap;
 
 public class Year {
   private int id;
-  private Map<Short, Month> months;
+  private Map<Integer, Month> months;
   
   public Year() {
     super();
@@ -15,7 +15,7 @@ public class Year {
      * sort themselves, if not, you cannot count on the being in the same order 
      * when comparing for testing purposes as they are only viewed as 
      * collections then. */
-    months = new TreeMap<Short, Month>();
+    months = new TreeMap<Integer, Month>();
   }
   
   public void setId(final int id) {
@@ -29,7 +29,7 @@ public class Year {
   /**
    * @return the indicated month
    */
-  public Month getMonth(final short monthId) {
+  public Month getMonth(final int monthId) {
     return months.get(monthId);
   }
